+++ App.tsx
@@ -499,6 +499,13 @@ export const App: React.FC<IProps> = (): React.ReactElement => {
                 </Button>
 
                 <FilterDrawer
+                  onRequestClose={() => {
+                    if (isFilterDrawerSecondaryViewActive) {
+                      setIsFilterDrawerSecondaryViewActive(false);
+                    } else {
+                      setIsFilterDrawerOpen(false);
+                    }
+                  }}
                   onApplyFilters={() => Alert.alert('Filters Applied!')}
                   onClearFilters={() => Alert.alert('Filters Cleared!')}
                   onClose={() => setIsFilterDrawerOpen(false)}