+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### v0.0.42
 
 - Fix Filter drawer back button does not close model (Android Specific)
+- Update Person Record Component, Fixes UI and Functionalities.
+- Add SelectPicker showCustomIcon Prop
+- Add style textAlignVertical:'top' to TextInput when multiline for same behavior in both platforms
 
 --
 
@@ -10,6 +13,7 @@
 
 - Add Note list item component
 - Add SelectPicker Required Prop
+- Add fontFamily on text Input
 
 --