+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.0.42
 
+- Fix Filter drawer back button does not close model (Android Specific)
 - Update Person Record Component, Fixes UI and Functionalities.
 - Add SelectPicker showCustomIcon Prop
 - Add style textAlignVertical:'top' to TextInput when multiline for same behavior in both platforms