+++ CHANGELOG.md
@@ -1,11 +1,21 @@
 # Changelog
 
+### v0.0.42
+
+- Fix Filter drawer back button does not close model (Android Specific)
+- Update Person Record Component, Fixes UI and Functionalities.
+- Add SelectPicker showCustomIcon Prop
+- Add style textAlignVertical:'top' to TextInput when multiline for same behavior in both platforms
+
+--
+
 ### v0.0.41
 
 - Add ReminderPicker Component
 - Add CalendarPicker Component
 - Add Note list item component
 - Add SelectPicker Required Prop
+- Add fontFamily on text Input
 
 --